% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khanetal2022im3gcamusa.R
\docType{package}
\name{khanetal2022im3gcamusa}
\alias{khanetal2022im3gcamusa}
\title{khanetal2022im3gcamusa}
\description{
\itemize{
\item Github: https://github.com/IMMM-SFA/khan-etal_2022_im3gcamusa
\item Webpage: https://khan-etal_2022_im3gcamusa.github.io/
}
}
